Many interwoven changes, again...
- Added config_bool(), which was already used in the previous commit :(
- get_series() now return (index, series_id), i.e. no longer the
`series` dict. (the caller has the means to extract it easily anyway
and it's not always needed.
- Merged refresh_all() and refresh_series() into refresh_series() that
takes a `subset` argument (a list of series id) tp refresh. It also
takes a `max_age` argument (in days) This also means that
cmd_refresh() needs to do a little more lifting with the
argument checking, as it should.
- Refreshing now uses the new "fancy" progress bar.