Added migration/fix for bad values for "archived" meta.
Currently, there's no DB meta data (e.g. a version number), so these
migration/fixes will be done at each load of the series db, regardless
if it's needed or not.
However, measurements show that this check (when there was nothing to fix) is
very fast: ~45µs for 150 series on my machine.