Added `calendar` command.
Prints N future (whole) weeks and for each day which episodes are
released then. By default, 1 whole week in the future up to the Sunday
that last week.

Where "whole" means e.g. if today is a Wednesday, and N = 1, it will print this
week and the next. Thus, N = 0 will print only this week (1 to 7 days).

Resolves #6.