Added filtering by director/writer/cast/years to `list` command
Renamed known_options -> command_options, and also restructured it to
be a dict (instead of a list), where the former `key` key is now the
key in that dict. This doesn't change much, other than the definition
takes less space.

Because of this, and some other recent related changes, the
options_help() function is no longer useful, we just use the
option_def() for this purpose as well.

Also renamed _0 -> _00 and _0fg -> _0, which pretty much changed half
of the source code :|