Implemented persistent, sequential list index.
Not sure why this wasn't the first impl though...

DB version bumped to 2.

'next_list_index' is stored in the db-level meta. Mostly so we don't
need to look it up when adding, but also when deleting the series with
highest index, that index doesn't get reused.

Also, `delete` command now prints the index of the deleted series.
Not really useful, but at least consistent with other commands.

The global --debug option now prints the DB version.

Resolves #10