The `unseen` command is now an alias for `show` (with opts)
Some options were added to show/ls for this to be possible, namely
-u (list only series with unseen episodes), -N (show next episode
only, no summary) and -f (make -e to also show future/unreleased episodes)

The commands `show` and `unseen` always had a pretty large overlap, code-wise,
so this is a good step to take.

For #13, this should be made into an actual macro (currently it's just
a bit of code).