Removed IMDb ID in list views
It's simply not needed.
`show -I` shows it (in fact, the entire IMDb link is shown)

Also, left-aligned the progress percentage on the title line, right
after the "ended" status label. This makes more sense, since the
progress is known/possible b/c the series has ended (or was cancelled).