Prefer python modules for compression, if available
priority order:
- zstandard module
- zstd binary
- lz4 binary (probably add module later)
- lzma module
- xz binary
- gzip module (in Python standard library)
- gzip binary

Superficial benchmarks show that using the zstandard module is about
similar in speed to using the external zstd binary. Maybe not the
awestriking improvement I was hoping for, but at least it feels like
the correct way to do it. ;)  I assume there's a slight reduction in
memory usage, at least.

Also, the db.py was getting cramped to all compression stuff was moved
to a separate file. It's all way too complex for this simple use
case. :)

Resolves #45