Added "expiry" of series data
If the data of an archived series isn't referenced in (currently) 30
days, it is removed.
Just the series data on disk, not the database entry.

Changed the migration to external data to only write non-archived
series. Also the refreshing of the series data will not do anything if
there's currently no data stored.

If any non-existing data is needed, it will be automatically downloaded.